+++ lib/core/crop/image_fit_mode.dart
@@ -1,0 +1,1 @@
+enum ImageFitMode { cropToFill, fitInside }