+++ lib/features/print_job/print_job.dart
@@ -4,3 +4,4 @@ export 'print_configuration_controller.dart';
 export 'print_configuration_screen.dart';
 export 'print_configuration_state.dart';
 export 'print_job_configuration.dart';
+export 'print_sheet_preview.dart';