+++ lib/platform/image_processing/image_processing.dart
@@ -1,0 +1,4 @@
+export 'dart_image_processor.dart';
+export 'image_processing_request.dart';
+export 'image_processor.dart';
+export 'processed_image.dart';